Disable `setuptools` management of dependencies
Typically `setuptools` will install our dependencies for us at build
time. As we are building our own package where we manage dependencies
ourselves, we don't need `setuptools` to do this for us. In fact it is
problematic for `setuptools` to install our dependencies at build time
as they may wind up in our Conda package, which would cause conflicts
and other problems.

To fix this, we tell `setuptools` to disable this behavior with the
`--single-version-externally-managed` flag. This requires a flag to
be added to log the files written. We really don't care about that
either, but `setuptools` won't let us skip it. So we add
`--record=record.txt` for this purpose.

ref: https://setuptools.readthedocs.io/en/latest/ez_setup.html#install-run-easy-install-or-old-style-installation