Raise serialization errors when spilling
The default behavior of `serialize` (and thus `serialize_bytes`) is to
capture any serialization errors as messages. The result is the error
gets raises on deserialization. However for spilling, this means that we
fail to spill and raise that error only after trying to restore that
data.  Instead go ahead and raise the error when trying to spill to
avoid thinking content has spilled successfully (when that is not the
case) and thus losing it.