Handle removal of `get_default_resource_type`
As `get_default_resource_type` was dropped in RMM recently, use the
newly introduce `rmm.mr.get_per_device_resource` instead to access the
resource on device `0` (configured to a unique device for each worker).
Since the memory resource itself is not realistically serializable,
instead grab the type of each resource to send back. This is all done
within a `lambda` to allow for a function that can be run on each
worker.