Fix versioneer (#448)
* Add `MANIFEST.in` to include versioneer components

Some of the components of versioneer may not have been included during
packaging. To fix this we add a `MANIFEST.in` file to ensure these
pieces are included.

* Always use versioneer

Make sure that we always query versioneer for the version. It already
handles cases where the package does not live in version control (like
being installed or in a tarball).