Recipe cleanup (#449)
* Drop `x.x` syntax

This has been deprecated for a while in Conda-Build and isn't strictly
needed any more. So go ahead and drop it.

* Replace `build` with `host`

Conda-Build 2 added the `host` environment for things that would be
needed to link against so as to distinguish it from `build` tools, whose
CLIs were only need (like CMake). Given this, go ahead and switch this
to `host` to match with that usage.