Add device memory limit argument to benchmarks (#683)
Adds a `--device-memory-limit` argument to the benchmark utils, to set a device memory limit on the instantiated `LocalCUDACluster`. This can be helpful when we want to force spilling to host (for context, I am doing this while benchmarking pack/unpack serialization).

Authors:
  - Charles Blackmon-Luca (https://github.com/charlesbluca)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)

URL: https://github.com/rapidsai/dask-cuda/pull/683