Add `--benchmark-json` option to all benchmarks (#700)
Makes `--benchmark-json` an option for all the benchmarks; also refactors the JSON dumping process so that the output file is a line-delimited JSON file, which can be appended to by subsequent runs of the benchmark script. This makes it a lot easier to process the output in Pandas with something like

```python
pd.read_json("benchmark.json", lines=True)
```

Also added the `device_memory_limit` option to the printed output of the scripts, so that it is visible if specified.

Authors:
  - Charles Blackmon-Luca (https://github.com/charlesbluca)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)

URL: https://github.com/rapidsai/dask-cuda/pull/700