Remove proxy object locks (#747)
This PR removes the thread locking of `ProxyObjects`, which greatly simplifies the avoidance of deadlocks. 

#### Motivation
Currently, both the `ProxyManager` and all `ProxyObjects` share the same lock, which prevents concurrent spilling and unspilling of `ProxyObjects`. This is a problem particularly if we want to implement [asynchronous memory spilling](https://github.com/rapidsai/dask-cuda/pull/744), which my plan is to implement next.

Authors:
  - Mads R. B. Kristensen (https://github.com/madsbk)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)

URL: https://github.com/rapidsai/dask-cuda/pull/747