Fix useless property (#944)
Make the`jit_unspill` property a local as discussed in #944. 

In addition, fix a bug that was introduced in `dask-cuda-worker` in #944: the `data` argument has to be a callable, and should presumably return an object that is unique for each nanny.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Peter Andreas Entschev (https://github.com/pentschev)

URL: https://github.com/rapidsai/dask-cuda/pull/944