Always use versioneer command classes in setup.py
The GIT_DESCRIBE_TAG and VERSION_SUFFIX environment variables are used
to control the name and version of the created conda/pypi package.
They should, however, not be used to control the version of the
installed package by overriding the versioneer cmdclass since that
leaves an unmodified _version.py file in the installed package
directory. A consequence is that the version reported by
dask_cuda.__version__ is "0+unknown". To fix this, always use the
versioneer-provided cmdclass.

While we're here, bring the conda version string into line with the
rest of the rapids ecosystem.