Override versioneer.get_versions when GIT_DESCRIBE_TAG is defined
Per PEP440, local version identifiers are disallowed for uploads to
public index servers. In CI, when we build pypi packages, we override
the versioneer-provided version with a PEP440-compatible version by
defining GIT_DESCRIBE_TAG and VERSION_SUFFIX. These are glued together
to produce a version which we then use to override the version
produced by versioneer.

This leaves us free to use the command classes provided by versioneer
in all circumstances. Closes #336.