Make local_directory a required argument for spilling impls
For automated cleanup when the cluster exits, the on-disk spilling
directory needs to live inside the relevant worker's local_directory.
Since we do not have a handle on the worker when constructing the
keyword arguments to DeviceHostFile or ProxifyHostFile, instead take
advantage of dask/distributed#7153 and request that we are called with
the worker_local_directory as an argument. Closes #1018.