Work-around for dask/distributed#7726 and dask-cuda#1148
The last atexit finalizer in the chain of finalizers that distributed
registers needs to be one in the weakref module. Otherwise we observe
hangs in the shutdown of UCX-Py where one thread has the gil but not
the UCX spinlock and vice-versa.

To ensure this, unregister the known shutdown finalizers in
distributed, add our weakref finalizer, and then re-register them.
Along with a final re-implementation of the "python shutting down"
handler which must run before other handlers in distributed.