Ensure plugin config can be passed from worker to client (#1212)
It is possible when printing a cluster configuration that attributes of a plugin are not serializable and so cannot be passed from worker to client. To guard against this, preemptively attempt to pickle the deduced configuration and just report an unknown configuration if that fails.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)

URL: https://github.com/rapidsai/dask-cuda/pull/1212