Allow disabling RMM in benchmarks
Allows disabling RMM in benchmarks via a new option `--disable-rmm`.
This change makes benchmarks a little more similar to RMM setup in
`LocalCUDACluster`/`dask cuda worker`, where not specifying
`rmm-pool-size` or specifying `None` as its value entirely disables
setting up RMM as the default allocator. Since for benchmarks it's
desired that the default is having an RMM pool we cannot change the
default `--rmm-pool-size` to `None` as that would make benchmarks run
much slower by default, therefore `--disable-rmm` is the closest we can
make this to the rest of Dask-CUDA.

Additionally add `--rmm-maximum-pool-size` for benchmarks.