Reorganize CI tests and benchmarks
Recently wheel tests have been failing, this is probably due to the
`DASK_CUDA_WAIT_WORKERS_MIN_TIMEOUT=20` configuration that is set for
conda tests but not for wheels. With this change, the tests and
benchmarks are now moved to new scripts that can be called by both conda
and wheels tests to provide unified configurations with sane values for
CI, and therefore should also resolve the CI failures.

Additionally removed `--durations` which was previously added to help
understand timeouts but has not been of much use since.