Make `dask.dataframe` optional (#1439)
Dask-CUDA currently requires that `dask.dataframe` be imported in a few places. We only do this to patch in explicit-comms shuffling and to register various dispatch functions. There is no fundamental reason that we need `dask.dataframe` to be installed if the user is not actually using `dask.dataframe`/`dask_cudf` in their workflow.

This PR essentially adds exception handling for "automatic" `dask.dataframe` imports (when `dask_cuda` is imported).

Authors:
  - Richard (Rick) Zamora (https://github.com/rjzamora)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)

URL: https://github.com/rapidsai/dask-cuda/pull/1439