Workaround hang with dask main and explicit comms
Dask>2025.3.0 will delay materialization of the task graph for more
operations, including expressions and HighLevelGraphs. The final task
graph will now be built on the scheduler.

The explicit-comms shuffle relied on the low-level graph being
materialized on the client. Parts of the task generation in
`ECShuffle._lower` use code that, apparently, doesn't run well on the
scheduler in the context of graph materialization.

This works around that issue by continuing to materialize the task
graph on the client before sending it to the scheduler. We only do this
when we're using explicit comms and the result to compute contains a
shuffle task.

https://github.com/rapidsai/dask-upstream-testing/issues/37#issuecomment-2779798670
contains more information on the issue.