Register OwnerWithCAI with dask.utils.sizeof (#1468)
Fixes the failing proxy test seen in https://github.com/rapidsai/dask-cuda/pull/1467.

---

The first change we see is in `get_device_memory_objcects.dispatch`


```python
import cudf
import dask_cuda.get_device_memory_objects
import dask.sizeof


def main():
    s = cudf.Series([1, 2, 3])
    
    items = dask_cuda.get_device_memory_objects.dispatch(s)
    print(items, dask.sizeof.sizeof(s), dask.sizeof.sizeof(items[0]))


if __name__ == "__main__":
    main()
```

Previously, that found the RMM DeviceBuffer. The sizeof the Series was 48 and the sizeof the DeviceBuffer was 24:

```
./old/bin/python ./debug.py 
[<rmm.pylibrmm.device_buffer.DeviceBuffer object at 0x7fb53dfd0940>] 48 24
```

Now, we get a `pylibcudf.column.OwnerWithCAI`. The sizeof the Series is still 48, but now the sizeof this OwnerWithCAI thing is 56, I think because Dask is using some default sizeof for unknown objects?

```
./new/bin/python ./debug.py 
[<pylibcudf.column.OwnerWithCAI object at 0x7fec2b6d5a00>] 48 56
```

This causes the test to fail because the program sees we've crossed some memory limit (thanks to 56 > 24) and spills the cupy array, which breaks the identity assert.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Richard (Rick) Zamora (https://github.com/rjzamora)
  - Peter Andreas Entschev (https://github.com/pentschev)

URL: https://github.com/rapidsai/dask-cuda/pull/1468