Add UCXX tests as separate nightly workflow (#1470)
Use a separate nightly workflow to launch UCXX tests. We have previously attempted to fork a job from the regular nightly workflow but that proved more challenging than it's worth the trouble, so now we switch to a cron-based job that is known to work and suffices for the current requirements.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/dask-cuda/pull/1470