Add separate `build-ucxx` workflow (#1472)
In #1471 the `build` workflow was made reusable but a separate `build-ucxx` workflow is necessary for `nightly-pipeline` because uploading conda/wheels build is not supported for a custom workflow. This change reverts making `build` reusable and use a new `build-ucxx` workflow only for the `conda-python-build` job.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)

URL: https://github.com/rapidsai/dask-cuda/pull/1472