Fixed flaky `test_cudf_cluster_device_spill` test (#1492)
The test fails because we're asserting that the worker's local data directory is supposed to be empty. There seems to be some a race condition between when the test thread checks the size of that directory and when whatever thread does the cleanup actually runs.

I think the intention of the test is to ensure that the `del cdf` and `gc.collect()` causes the worker memory (and files spilled to disk) to be cleaned up on the worker. However `__del__` is synchronous and the cleanup is happening asynchronously on the worker. `__del__` just emits the RPC instructing the worker to free the memory and clean up the files, but doesn't block until that happens.

This structure updates the test to run the check in a loop and only fail after the check fails to pass after some timeout duration (3 seconds). With these changes the test passes 100 times using `pytest-repeat`. Previously it usually failed within 5-ish repeats.

While not strictly necessary, I've restructured this part of the test a bit. Previously, we used `client.run` to run a function that collected some data on the worker and made some assertions about it (in this case, that the worker data was empty). I've restructured the test to just collect the data on the worker and make the assertion on the client. This makes the tracebacks emitted by pytest nicer and aids in debugging since you can see a bit about why the test failed. We can consider doing this in more places in the tests, but IMO it's not a high priority.

Closes #1490

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)

URL: https://github.com/rapidsai/dask-cuda/pull/1492