Get all workers from scheduler info
For no good reason other than making Jupyter pretty printing nicer,
https://github.com/dask/distributed/pull/9045 broke the API of
`client.scheduler_info()` by changing the default number of workers that
are retrieved to `5`.

We use this information in various places to ensure the correct number
of workers have connected to the scheduler, and thus this change now
specifies `n_workers=-1` to get all workers instead of just the first
`5`. This happened not to be seen in CI previously because we don't run
multi-GPU tests, let alone with more than `5` workers. Running various
tests from `test_dask_cuda_worker.py` on a system with 8 GPUs will cause
them to timeout while waiting for the workers to connect.