Test rapids-dask-dependency UCX protocol selection (#1520)
Add a test for rapids-dask-dependency UCX protocol selection based on installed packages. The test was initially intended for rapids-dask-dependency but since there are no GPU tests in there it was instead moved here. This a temporary test that will be removed in the scope of #1517 .

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/dask-cuda/pull/1520