Fixes for `test_worker_force_spill_to_disk` (#1562)
Two fixes for this test, which hopefully preserve the cases it intends to cover:

1. Generate data on the worker, rather than on the client and serializing
2. Ensure that the coroutine making the assertions is awaited

This brings the test runtime down from 33s to 10s locally.

Spotted while looking at #1561. It shouldn't have an effect on that one way or the other.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)

URL: https://github.com/rapidsai/dask-cuda/pull/1562